// apps/edu-platform/src/models/LessonCollection.ts
export type LessonCollection =
  | 'lessonsOfe'
  | 'livesOfe'
  | 'lessonsCursor'
  | 'lessonsOtsCore'
  | 'lessonsOtsReact'
  | 'lessons10xDevs1'
  | 'lessons10xDevs2'
  | 'lessons10xDevs2EN'
  | 'lessons10xDevs3'
  | 'lessons10xDevs3En'
  | 'lessons10xDevs3Prework'
  | 'lessons10xDevs3PreworkEn';